requirements.txt

click==4.0
ecdsa==0.13
Fabric==1.10.1
Flask==0.10.1
Flask-FlatPages==0.6
Frozen-Flask==0.11
itsdangerous==0.24
Jinja2==2.7.3
Markdown==2.6.1
MarkupSafe==0.23
paramiko==1.15.2
pycrypto==2.6.1
PyYAML==3.11
-e git+git@github.com:chipperdoodles/3color.git@50300a74eee87cd97e1b3c238759c2a01c37b06c#egg=threecolor-master
Werkzeug==0.10.4
